script: Move SVG DOM interfaces to `script/dom/svg/`
