+++ components/script/dom/svg/svgimageelement.rs
@@ -14,7 +14,7 @@ use crate::dom::bindings::str::DOMString;
 use crate::dom::document::Document;
 use crate::dom::element::AttributeMutation;
 use crate::dom::node::{Node, NodeTraits};
-use crate::dom::svggraphicselement::SVGGraphicsElement;
+use crate::dom::svg::svggraphicselement::SVGGraphicsElement;
 use crate::dom::virtualmethods::VirtualMethods;
 use crate::script_runtime::CanGc;
